Update patch set 18
Patch Set 18:

(8 comments)

> Patch Set 17:
> 
> (7 comments)
> 
> Does exception safety matter here? That's problematic without the make_XXX functions.
> 
> I would generally try to stick to either sp<> or std:: methods and avoid invention. It seems really easy to get this badly wrong.

Yes, I would like to use sp<> here, but I can't because it depends on things outside of the NDK. Re invention, see comments in-line. Certainly if we can make a librefbase, a lot of this would be cleaner.

Patch-set: 18
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>