Update patch set 5
Patch Set 5:

I should mention there are two other ways I could solve this problem that I can think of:

1). having multiple APIs. AParcel_readStringSize and AParcel_readStringWithSize

I didn't like this because it opens up the possibility of the client to call the sizes out of order.

2). having the client allocate a buffer and call read multiple times if it needs more data (like POSIX read) (just discussed w/ elsk@)

But this requires guessing the size in advance and possible re-allocations.

Patch-set: 5
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>