Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This isn't very clean, but it is accurate. What do you think of the general idea (seeing the changes to the test)? If you like it, I can try cleaning this up more.
> 
> I like it, just worried about breaking stuff that (wrongly) assumes the current behavior. Have you run this on a Pixel to see how often these paths are hit?

Thinking about it more, I'm not sure why EX_TRANSACTION_FAILED exists. If it's supposed to denote a binder transaction failure, that means it failed at a low-level (kernel or native userspace), and we use FAILED; why would we need an Exception?

Is it just a hack to indicate that if mErrorCode is set, mException should also be set to something that is not an error?

Patch-set: 1
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>