Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > This isn't very clean, but it is accurate. What do you think of the general idea (seeing the changes to the test)? If you like it, I can try cleaning this up more.
> > 
> > I like it, just worried about breaking stuff that (wrongly) assumes the current behavior. Have you run this on a Pixel to see how often these paths are hit?
> 
> Thinking about it more, I'm not sure why EX_TRANSACTION_FAILED exists. If it's supposed to denote a binder transaction failure, that means it failed at a low-level (kernel or native userspace), and we use FAILED; why would we need an Exception?
> 
> Is it just a hack to indicate that if mErrorCode is set, mException should also be set to something that is not an error?

mErrorCode is overloaded. If the exception code is EX_SERVICE_SPECIFIC, then it is a service error code. If the exception code is EX_TRANSACTION_FAILED, then it is a low level error (and mErrorCode is a status_t). As far as I can tell, after this change, if mErrorCode is non-zero, then mException is either EX_SERVICE_SPECIFIC or EX_TRANSACTION_FAILED.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>