WIP: libbinder_ndk: read/write array for primitive types.
There are two types of primitives to be considered:
1) those which are contiguously stored in arrays
2) those which are not contiguously stored in arrays (only bool)

This allows (1) to be read contiguously (however, right now, this
implementation does not read them contiguously at once and neither does
libbinder. The current API allows this to be optimized.

For bool, since things like std::vector store the bool vector in a
packed way, this API allows for that.

This is a bit messy. Possible other directions:
- read all types like bool is being read (can't memcpy in the future)
- don't allow vec<bool> (non-option)

Bug: 111445392
Test: WIP
Change-Id: I3934f7d5cd749810c4a512ac3fb57b63f9ae4ea5