Update patch set 1
Patch Set 1:

I'm still not a fan of enabling this generically; there may be scenarios where you just don't care to read an entire Parcel, and manually having to wind to the end in those cases is just weird. Can we make this opt-in somehow, as something we only do in cases where we know we want this behavior?

Related, how often does this actually trigger? Is the hybrid interface one the only instance you found?

Patch-set: 1
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>