Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > I'm still not a fan of enabling this generically; there may be scenarios where you just don't care to read an entire Parcel, and manually having to wind to the end in those cases is just weird. Can we make this opt-in somehow, as something we only do in cases where we know we want this behavior?
> > > > 
> > > > Related, how often does this actually trigger? Is the hybrid interface one the only instance you found?
> > > 
> > > I haven't seen anything else trigger it. After I figure out why HybridInterface still isn't reading all of the data, I was planning on cleaning this up and merging it.
> > > 
> > > I don't think there's any case where someone would want to regularly not read all of the data. If it's not important to read, why even write it?
> > 
> > I believe you owe me some beer then, because we took a bet this would uncover >5 (I believe) real bugs? :-)
> > 
> > Whether it's important to read depends on the state of the reader. Simple example: the writer can write some state information about an object with id X. So the Parcel contains id X, followed by some information about the object with id X. When the reader receives it, it reads the ID, sees that it is X, and then decides it's not important to know about X anymore - maybe one if its own clients has told it it's no longer interested in X.
> > 
> > I know in cases of auto-generated code, you would normally read all of the Parcel by design, and in those cases I don't mind checking for this (although it's also somewhat annoying for the auto-gen code because it has to rewind to the end in case failure) - and I still worry about backwards compatibility in those cases.
> > 
> > But aside from that, there is no contract in binder that says you have to read everything you write, and I don't see why there should be.
> 
> Sure :) For simplicity/security/sanity, I think it makes sense for a reader to validate the whole transaction. For versioned structs in stable AIDL, we should know the entire size of the full struct. For transactions that need to be optimized not to read certain content, they should be optimized to also not write that unimportant stuff. You are fully against adding a warning here then?

My point is that the writer can't be aware what the state of the reader is - this is IPC and they are decoupled. So it's unrealistic to expect the writer to only write what the reader needs. The thing is, for our auto-generated interface code I agree with you - the reader should read and validate everything, because that is what the auto-generated code is designed to do (and the receiver of the data can then still decide to drop it if it doesn't care). But this is libbinder, a Parcel is an abstract container for IPC'd data that can be transacted. libbinder shouldn't care what you do with that data - it's a transport.

Besides that, I also don't see any real upsides to this; binder has been around for ages, yet the only instance where we found the Parcel wasn't read entirely was a case where the descriptor wasn't read (which arguably didn't really matter anyway)? I'm reluctant to add even the minor overhead of a method call + condition to every binder transaction for something like that.

Patch-set: 1
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>