surfaceflinger: Fix hwcomposer command errors
When using the hwcomposer API, there are two paths that the command may
take. One path results in a direct synchronous IPC message, whereas the
other path will buffer the commands and dispatch later just before
present.

This opens the possibility for out-of order execution of those commands.
and can result in the following errors showing in logcat:

    HwcComposer: command 0x3010000 generated error 3

One scenario where this happens is the destroyLayer() command. This
command is dispatched immediately, but it's possible that the buffered
command stream [that has not yet been sent] contains commands that
reference that layer.