Update patch set 1
Patch Set 1:

Dan, Chia-l,

I've now made a fix for this issue in our hwcomposer codebase; to ensure that the empty-list scenario presents a black screen to display.

So this patch for surfaceflinger is not needed for us. I'll let me guys decide if you still think its useful for surfaceflinger and want to submit it.

Patch-set: 1
Reviewer: Gerrit User 1044843 <1044843@85c56323-6fa9-3386-8a01-6480fb634889>