surfaceflinger: Fix EGL deadlock during composition
This patch fixes a deadlock during GL/EGL rendering, due to the how the
FramebufferSurface latching mechanism has been changed in the transition
to HWC2.

The FramebufferSurface is responsible for consuming the buffers from
composition rendering, and uses a latching mechanism. This behaves in a
way that there is always 1 buffer acquired out of the bufferqueue, which
is only released back to the bufferqueue once the next replacement
buffer has been successfully acquired. This means there is a period of
time when 2 buffers are acquired out of the queue. With a
double-buffered bufferqueue, surfaceflinger needs to ensure that no
GL/EGL commands are called during this time when both buffers are
acquired, otherwise deadlock could occur within the GL/EGL driver.

Previously with the HWC1 codepaths, surfaceflinger ensured this because
FramebufferSurface would call acquireBuffer and releaseBuffer in
sequence and within the same function.

However with the new HWC2 codepaths, acquireBuffer and releaseBuffer are
now called at different points during composition. And during that time,
surfaceflinger will also do composition work for other displays.  This
means there is a large window of time, during which 2 buffers are
acquired out of the queue. If any GL/EGL calls are triggered during that
time, its likely that they will get stuck trying to dequeue a free
buffer from the queue.

This issue has been seen during virtual display usecases. During
compsition, surfaceflinger first works on display 0, and has 2 buffers
acquired out of the FramebufferSurface. Next it does the work for
composition of virtual display - if this fails for some unrelated
reason, and during failure handling, surfaceflinger calls eglMakeCurrent
on display 0. Internally EGL/GL will then try to dequeue a free buffer,
but get stuck waiting because surfaceflinger has both buffers acquired.

This patch updates the points in the composition loop, which trigger the
calls to acquireBuffer and releaseBuffer, so that they are done much
closer together and reduce the possibility of surfaceflinger making any
GL/EGL calls during that time.