InputFlinger: Support amplitude control of input device vibrators
In N, amplitude control was added for Vibrator class and supported
only by the SystemVibrator. The existing implementation for
InputDeviceVibrator is only added when the input device supports
the FF_RUMBLE force feedback event. FF_RUMBLE has parameters for
controlling rumble amplitude.

Most modern gamepads contain multiple rumble motors, generally
a "strong" and "weak" rumble motor that provide a low frequency,
high amplitude rumble effect and a high frequency, low amplitude
rumble effect respectively.

Changes:

Add plumbing to get framework controlled rumble amplitude to
InputReader and EventHub.

Fix bug with vibration API. In N, the VibrationEffect method which
takes both timings and amplitudes states that each timing/amplitude
pair makes the vibration last for the specified duration with the
specified amplitude, 0 meaning the motor is off. If this method
is used with an input device vibrator, it turns off for every even
numbered element, regardless of the amplitude value. The method
which takes only timing values works as expected. This problem
was not fixable without sending the amplitude to the input device
vibrator in the first place, so this patch allows this to be fixed.

Test: Connect a DualShock4, create VibrationEffect objects with
      a pair of variable amplitudes, and call Java API vibrate()
      with the objects.
      Example:
          vibrator.vibrate(VibrationEffect.createOneShot(
                  VibrationElement.create(5000L,
                          new int[] {largeMotor, smallMotor})));
Bug: 38511270
Change-Id: I70aa067001b8756786dcd591b7e448fee9b69b9d