Enable installd to use ART binaries from the Android Runtime APEX.
This is only true if the product includes the Android Runtime APEX,
i.e. if the product variable `dont_include_runtime_apex` is not set;
otherwise, installd keeps using ART binaries from /system.

This is a temporary change that will be altered to only use the
Android Runtime APEX when all products ship it.

Test: Device boot test with Android Runtime APEX.
Test: Device boot test without Android Runtime APEX.
Bug: 113373927
Change-Id: I79652a3b3fbdd5761cf81be56c524cce0dab2ca5