Update patch set 1
Patch Set 1: Verified+1

The follow-up for this CL would be to do this for every other type (make them all 'nullable' like in Java at the NDK layer, and make the optionality of the nullability present only at the AIDL layer). This is as discussed on the backend design doc, based on discussion w/ jreck.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1