Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> It looks like isCreatedFromMainThread isn't actually used any where?

Yes, isCreatedFromMainThread is actually used in the ContainerLayer, because SF only creates ContainerLayer in the main thread at present. I do not know whether SF has other requirements which use the main thread to create BufferLayer or ColorLayer in the future, so I modify their constructor in this patch. Then their constructors can have the same interface with each other. If SF will create them in the main thread, we do not need to update interface again.

Patch-set: 5
CC: Gerrit User 1395460 <1395460@85c56323-6fa9-3386-8a01-6480fb634889>