NDK accepts non-user commands
NDK accepts non-user commands (i.e. command whose ID is greater than
LAST_CALL_TRANSACTION). This is required because the AIDL compiler
generates some non-user commands like getInterfaceVersion which is
handled by the generated code and not by the binder library.

When the non-user command is not handled by the binder_ndk, then it
falls back to the binder.

Bug: 110760419
Test: atest under cts/tests/tests/binder_ndk
Change-Id: I5e76ec4a9e68121bbdff8edb29f08a3a62fc16b5