Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Can we move these commands to be at the end of the range of user commands then? This is exposing a whole separate set of functions that aren't defined in the NDK. If AIDL wants to define new functions, I think it should be within the range of user commands. It's at a higher level than libbinder/libbinder_ndk.

Patch-set: 2
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1