Update patch set 1
Patch Set 1:

> What's the process for getting this flipped?

When this is ready to ship and the Q branch has been created, revert this in the Q branch. That'll make it available in the Q sysroot artifact and when I'm ready to ship Q APIs in the NDK I'll use that one.

Gerrit doesn't have any rules for requiring API council approval on that change (I'm not sure if that rule can even be written properly), but make sure we get added just so we know what's going on.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>