Update patch set 1
Patch Set 1:

(4 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> Why are we separating FULL from DEFAULT? I mean, what will DEFAULT do?

Thanks for the review comments! I'll resolve them!

I think Nandana has explained the underlying motivation in creating the DEAFULT mode in detail. The below is my understanding of adding the DEFAULT mode.

We are here trying to segregate the concept of FULL bugreport as it is a system server call, with a few predefined extra options in the base call. We can add or remove certain options as and when required in the future. 

DEFAULT is a mode with no extra options or mode already set.

Patch-set: 1
Reviewer: Gerrit User 1404253 <1404253@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0