Update patch set 5
Patch Set 5:

(5 comments)

Just a couple things from my end.  I think the only one that needs fixing in this CL (as opposed to followup changes) is the data directory one.

It might also be good to come back into this code and find a way to remove the unhealthy relationship between dumpstated and the shell app.  I don't think it's super urgent, but we're starting to accrete a lot of legacy ways of calling this thing for each of the iterations.  The logic around figuring out which file we're writing to seems to be reaching the limit of reasonable complexity.  Are there any tests for that part of the code?

Patch-set: 5
Reviewer: Gerrit User 1000701 <1000701@85c56323-6fa9-3386-8a01-6480fb634889>