Update patch set 4
Patch Set 4:

> What do you think?

I still think the "default" and "full" options should be the same. I agree that ideally "default" would be the "bare minimum" option required to take a bugreport, but once we reach that point, there should be no option (as a string defining the bugreport type) at all, but rather flags in the request that shape the behavior (so we could haven dozen of different options, based on combination of flags).

Besides, now we need to explicitly set a property (dumpstate.options=bugreportfull) when taking a bugreport using POWER-UP-DOWN, as that option simplify triggers the bugreport service, that calls dumpstate without passing any system property - how are we keeping that behavior with this change?

Patch-set: 4
Reviewer: Gerrit User 1013984 <1013984@85c56323-6fa9-3386-8a01-6480fb634889>