Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > What do you think?
> 
> I still think the "default" and "full" options should be the same. I agree that ideally "default" would be the "bare minimum" option required to take a bugreport, but once we reach that point, there should be no option (as a string defining the bugreport type) at all, but rather flags in the request that shape the behavior (so we could haven dozen of different options, based on combination of flags).

I think Dumpstate is not in a state where its flags can be exposed to clients. There are some flags that make sense to clients and some that don't. Hence the modes (which already existed). I agree at some point when the options are cleaned up it makes sense to offer clients the ability to choose flags themselves. Like you said there are too many combinations possible. Also the modes we offer are a little ad hoc. 

> 
> Besides, now we need to explicitly set a property (dumpstate.options=bugreportfull) when taking a bugreport using POWER-UP-DOWN, as that option simplify triggers the bugreport service, that calls dumpstate without passing any system property - how are we keeping that behavior with this change?

It will still work just as before though, because the commandline arguments have not been removed. But it will not call itself a "full bugreport" in extra_options. Which seems ok to me.

Patch-set: 4
Reviewer: Gerrit User 1336554 <1336554@85c56323-6fa9-3386-8a01-6480fb634889>