Revert^2 "Have `otapreopt_chroot` mount APEX packages in the chroot dir."
We need at least the Android Runtime APEX to be able to run `dex2oat`
in `otapreopt`. For simplicity, reuse `apexd`'s logic and mount all
the APEX packages found on the system partition in /postinstall/apex.

This is a temporary change. The final design will integrate A/B OTA
updates and `apexd`.

This reverts commit 00b5b3f01b7b6bf96d0d66d650ede8aa4734bf3f.

Change-Id: Ifa45c282d04ae443e55bd41d8c9cbb4ea5f837d4
Test: m otapreopt_chroot
Test: A/B OTA update test (asit/dexoptota/self_full).
Bug: 113373927
Bug: 120796514