vulkan: Update Vulkan headers to 1.1.96
Key changes since 1.1.94:
* Corrected definition of VkPhysicalDevicePCIBusInfoPropertiesEXT

This is a rare ABI-breaking change. The Vulkan Working Group decided to
do it this way -- rather than deprecate the extension and create a new
one with the correct definition -- because no GPU vendor had made an
official driver release that supported the extension yet (it was present
in some open-source drivers, but hadn't been included in one of their
regular numbered/official releases).

Test: build
Change-Id: I085cf6ec17d8d6146299663689cf3b20fa4200e8