Detach android InputDevice ID from EventHub ID
Some physical devices contain more functions than can be addressed by
a single Linux evdev device. Examples of such devices are the Sony
DualShock 4 and Wacom Tablets, which appear as multiple evdev devices
sharing the same value for the EVIOCGUNIQ ioctl. As more instances of
such devices hit the market, apps need a way of figuring out which
InputDevices are part of the same physical device.

Per conversation with the android input team, a solution proposed to
this problem is to merge multiple EventHub devices (which have a 1:1
relation with evdev) into a single android InputDevice.

This patch addresses the first half of this problem, which is to
detach the EventHub ID from the InputDevice ID.

Changes:

Added enum value 'END_RESERVED_ID' to EventHub to represent the first
available InputDevice ID for random assignment. The android framework
reserves two EventHub/InputDevice IDs that must not change - one
for the virtual keyboard and one for the built-in (hardware) keyboard.

InputMappers are associated with an EventHub ID ('eventId') as well
as the InputDevice ID ('deviceId'). eventId is used to query EventHub
and deviceId is used for reporting events to the android framework.

All methods which wrapped EventHub access now take an eventId, rather
than the deviceId.

InputReader maintains an additional KeyedVector to relate eventIds to
InputDevices.

Test: atest inputflinger_tests
Test: see logs for new device messages where eventIds don't necessarily
      match deviceIds

Change-Id: Ib7c0ae15a6ca4216b40e82215adc6504827161cc