Support multiple EventHub devices per InputDevice
Some physical devices contain more functions than can be addressed by
a single Linux evdev device. Examples of such devices are the Sony
DualShock 4 and Wacom Tablets, which appear as multiple evdev devices
sharing the same value for the EVIOCGUNIQ ioctl. As more instances of
such devices hit the market, apps need a way of figuring out which
InputDevices are part of the same physical device.

Per conversation with the android input team, a solution proposed to
this problem is to merge multiple EventHub devices (which have a 1:1
relation with evdev) into a single android InputDevice.
This patch addresses the first half of this problem, which is to
detach the EventHub ID from the InputDevice ID.

This patch addresses the second half of this problem, which is to
merge EventHub devices which share the same EVIOCGUNIQ value into
a singular android InputDevice.

Changes:

Only create a new InputDevice if no input device matches a new
EventHub device's descriptor.

Forward InputDevice controls to all EventHub devices associated
with that InputDevice.

Rebuild sources and classes as EventHub devices are being added.

Test: atest inputflinger_tests
Test: plug in a DualShock 4 to an android device and observe that
      touchpad events come from the same InputDevice ID as gamepad
      events

Change-Id: I39d29699739882d1b85c1ba13d06007cea259b4b