Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > > Patch Set 7: Code-Review+1
> > > 
> > > I don't entirely understand this change because of the large number of points where we're manipulating filesystem permissions. It makes me a bit nervous but I can't see anything obviously wrong.
> > > 
> > > Do we need to worry about race conditions in this change? Specifically, is it possible for the application author to create a file in their home directory, and have installd "upgrade" it to executable? Either via symlink attacks, or renaming the files while the restorecon is occurring?
> > > 
> > > LGTM
> > 
> > Hmm, you're right. There is a race there. We would need an variant of selinux_android_restorecon_pkgdir which accepts fds, or have the oat inherit something else than app_data_file...
> 
> selinux_android_restorecon_pkgdir is a wrapper around setfilecon() which parses the seapp_contexts to determine the appropriate file label. However, if you have some other mechanism to determine the label, fsetfilecon() is the version which sets an SELinux label on a file descriptor.
> 
> * https://linux.die.net/man/3/setfilecon

Rewriting what I wrote because it doesn't make any sense.

selinux_android_restorecon_pkgdir is a wrapper around setfilecon(). The function selinux_android_restorecon_pkgdir parses the seapp_contexts to determine the appropriate file label, passing that file label to the setfilecon() function. However, if you have some other mechanism to determine the label, fsetfilecon() is the version which sets an SELinux label on a file descriptor. So perhaps it makes sense to have a function which returns the executable label for a particular package without changing that label directly? Or, like you said, a function which takes a file descriptor instead of a filename.

Patch-set: 7
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>