Update patch set 1
Patch Set 1:

We should always log to logcat, otherwise we might not able to see the message again later  (for example, if the user cancels the bugreport using the notification and we delete it).

I believe the reason we have 'log only' is for the cases where we're running on dry_run mode, so we can still measure the duration but don't need to effectively log the message in the bugreport text file.

-- Sent from my phone

Patch-set: 1
Tag: mailMessageId=<CAA6s5yPHW_=d1fxz=7x6TzLq5smU0r3RtYA05ndt53QfxX2k=g@mail.gmail.com>