Update patch set 1
Patch Set 1:

Why is this so complicated? Why does installd have to know about this, at all?

Alternative: add a mode for dex2oat that *only* loads the image? We already have cutouts for "no dex file fallback" and others. That would keep the intricacies of the ART module in ART.

Patch-set: 1
CC: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>