Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why is this so complicated? Why does installd have to know about this, at all?

It's because it's spanning two projects that you characterize this as complicated? We already went with clearing BOOTCLASSPATH in installd for not making ART complicated, so this is just an alternative, and cleaner way as now being explicit what the boot classpath is.

> 
> Alternative: add a mode for dex2oat that *only* loads the image? We already have cutouts for "no dex file fallback" and others. That would keep the intricacies of the ART module in ART.

Vladimir had the valid observation that without more woops, this would break our run-test device testing. So we'd move the complexity here. Also, not relying on env vars and being explicit does simplify the debugging.

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0