Fix sourceCrop transformation when device uses landscape LCM.
Testcases capture the entire screen, so sourceCrop is passed portrait device resolution.(0,0 – 1600,2560).
When the device uses landscape LCM, in fact, the size value stored in the display device contains the rotated value and this rect is set in the dispScissor. (0,0 – 2560,1600).
The code implemented inside renderScreenImplLocked makes sourceCrop equal to dispScissor, but there is a problem in processing the rotation transform in sourceCrop according to hardware rotation.

Bug:
Test: run cts-on-gsi -o --include-filter "CtsAssistTestCases android.assist.cts.ScreenshotTest#testBlueScreenshot" --include-filter "CtsAssistTestCases android.assist.cts.ScreenshotTest#testGreenScreenshot" --include-filter "CtsAssistTestCases android.assist.cts.ScreenshotTest#testRedScreenshot"

Change-Id: I3724a0c612981149977493c3bdc3d18f469767c5
Signed-off-by: Mark Hong <sungmin.h@samsung.com>