libbinder_ndk: unlinkToDeath dead proxy
When you call unlinkToDeath on a dead proxy, libbinder returns
DEAD_OBJECT. This behavior is duplicated in libbinder_ndk.

This CL fixes libbinder_ndk so that you can still unlinkToDeath after
a binder object dies, and it documents when DEAD_OBJECT is expected.

However, this also leads us to a situation where unlinking to death
is needed (to clear one of many obituaries after onDied is called
when a death recipient is used with multiple binders). In this case,
unlinkToDeath returns DEAD_OBJECT, but the call is still required
to clear the binder from the death recipient list.

Bug: N/A
Test: ./runtests.sh
Change-Id: Ib092b25273a51f5afb12904bfe491927ad404f8d