Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> LGTM
> 
> I just wonder if this behavior is consistent with the Java/C++ counterpart.

It is. In libbinder, BpBinder holds all of the state here though, so DeathRecipient doesn't need to clean it up. However, we are holding the cookie here to pass back to the client itself, so we need to do a bit of extra work (although binder doesn't do this?).

I'll look at it with fresh eyes on Monday. Any thoughts Martijn?

Patch-set: 5
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>