APP may display wrong direction in landscape LCM
When PRIMARY_DISPLAY_ORIENTATION is 90, 180 or 270, there are two issues as below.
1. APP preview displays in wrong direction
2. When APP is launched, APP may display in landscape mode firstly and then change
to portrait mode

The root cause is WMS/AMS use captureLayers instead of captureScreen, the captured
layer displays in wrong direction with surfaceflinger correction.

As captureLayers's CaptureFill is CLEAR, captureScreen's CaptureFill is default
OPAQUE, we can use ReaderArea CaptureFill to separate captureLayers and
captureScreen. And SurfaceFlinger disables the correction if cpatureLayers has been
called.

Bug: 129369370
Test:
1. APP preview displays in correct direction
2. APP is launched normally
3. The  direction of captrueScreen is correct
Change-Id: I52556a1397af69dd1b7186d5f2981cd1751e2add