Remove screenshot_fd as optional param
Binder does not allow invalid fd to be passed in function calls, so
remove "optional" in IDumpstate.aidl

Also, remove TODO to make screenshot fd optional, as the binder call
requires valid screenshot fd.

Also, we do not need valid screenshot fd for lite bugreports
(wifi/telephony/interactive etc). Although the Binder makes sure that
valid fds are passed, this is just an additional check on top of that.

Bug: 128981582
Test: Works as expected, just did a bit of clean up.
Change-Id: Idf6093832fea0f11f12ff2bc0fd1dcec6fd502df