Add netd to the list of native processes dumped for bugreports
This adds netd to the list of important native processes that
we want the stack traces of when collecting a bug report.

Out of 1:15 of dump, taking the stack traces of all these
processes is taking about 2.5 seconds on crosshatch according
to my (low data count) measurements. Adding netd seems to
increase the average by about 1 second measured over 2x5 runs.

Test: manual
Bug: 128804277
Change-Id: I2a560af11c2b11cdae895c68ae9db6a9d148f037