Refactor VelocityTracker_test
The test will now support full touch gestures (DOWN, POINTER_DOWN, MOVE,
etc).
Add a disabled test for three-finger tap (test currently fails, fix will
be done in a separate CL).

Add per-strategy velocity checks. This will eliminate redundant testing
and will ensure that both strategies are working at a given time.

Because of the above, some of the values for impulse velocity had to be
revised down. This is due to the change of the impulse strategy
implementation in ag/2837875, where we changed the initial condition
for the equation, which led to lower overall velocity values,
but we never adjusted the values in this test.

Bug: 129797752
Test: atest libinput_tests
Change-Id: Ife0a572166fcfd3c5135937fe8c46d67af7625b3
Merged-In: Ife0a572166fcfd3c5135937fe8c46d67af7625b3