SF: Add an rounding behaviour in computeGeometry.
When rendering client layers, need to call clearWithOpenGL
is used to clear the hardware layer's render area, need to
add round option in order to make the layer's render area
which is same as this layer's hwc frame. Otherwise, there
may be a shift between GPU_TATGET and hardware layers.

Test: build pass
BUG: 130838944

Change-Id: I1ac25f9438e2ae89390f63767732d6b4d4f2cede
Signed-off-by: wukui1 <wukui1@xiaomi.com>