dumpstate: Allow GetPidByName() to fail gently
If a pid can't be found, this might not rise to the level of a
logcat ERROR, depending on what the caller is expecting.  So we
allow the caller to optionally pass in a log level severity for
these cases.

We switch our search for the pid of "/system/bin/anrd" to an INFO,
since many devices do not have this binary.

Bug: 132919954
Test: TreeHugger
Change-Id: I287e622dc72b8fcb30562a811af33136b8b11f80